Add files via upload
In this game, you play as a ninja who runs to save the lives of her clan. The player jumps to avoid obstacles in her path, such as cats and rocks. She can also throw shurikens which allow her to eliminate the arrows of enemies who try to stop her. The goal is to run as far as possible; the player loses when the character comes into contact with an obstacle.